Replace wrappers with extensions
